Update Software Application
Changes done in the file to have a corrective ordering and error fixing in the library as in working with the code to manipulate the the menu driven program of the file to be executed as in the the code was not in the intendation of running the file properly by represention in CLI and GUI Version.In this program, you can instruct to OS using Python. This program gives the idea of how can we make a rule-based box.